Moved a lot of stuff around, improved parselet loading
